refactor: split up instance.py into individual files with tools
