+++ awslabs/rds_management_mcp_server/tools/db_instance/create_instance.py
@@ -1,0 +1,234 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Tool to create a new Amazon RDS database instance."""
+
+import asyncio
+from typing import Any, Dict, List, Optional
+from loguru import logger
+from mcp.server.fastmcp import Context
+from pydantic import Field
+from typing_extensions import Annotated
+
+from ...common.connection import RDSConnectionManager
+from ...common.decorator import handle_exceptions
+from ...common.server import mcp
+from ...common.utils import (
+    add_mcp_tags,
+    check_readonly_mode,
+    format_aws_response,
+    format_instance_info,
+    get_engine_port,
+    validate_db_identifier,
+)
+from ...constants import (
+    ERROR_INVALID_PARAMS,
+    ERROR_READONLY_MODE,
+    SUCCESS_CREATED,
+)
+
+
+CREATE_INSTANCE_TOOL_DESCRIPTION = """Create a new Amazon RDS database instance.
+
+This tool creates a new RDS database instance with the specified configuration.
+You can create a standalone instance or add an instance to an existing DB cluster.
+
+<warning>
+This operation creates AWS resources that will incur costs on your AWS account.
+</warning>
+"""
+
+
+@mcp.tool(
+    name='CreateDBInstance',
+    description=CREATE_INSTANCE_TOOL_DESCRIPTION,
+)
+@handle_exceptions
+async def create_db_instance(
+    db_instance_identifier: Annotated[
+        str, Field(description='The identifier for the DB instance')
+    ],
+    db_instance_class: Annotated[
+        str, Field(description='The compute and memory capacity of the DB instance, e.g., db.m5.large')
+    ],
+    engine: Annotated[
+        str, Field(description='The name of the database engine to be used for this instance')
+    ],
+    allocated_storage: Annotated[
+        Optional[int], Field(description='The amount of storage (in GiB) to be allocated for the DB instance')
+    ] = None,
+    master_username: Annotated[
+        Optional[str], Field(description='The name of the master user for the DB instance')
+    ] = None,
+    master_user_password: Annotated[
+        Optional[str], Field(description='The password for the master user')
+    ] = None,
+    manage_master_user_password: Annotated[
+        Optional[bool], Field(description='Specifies whether to manage the master user password with AWS Secrets Manager')
+    ] = True,
+    db_name: Annotated[
+        Optional[str], Field(description='The name for your database')
+    ] = None,
+    db_cluster_identifier: Annotated[
+        Optional[str], Field(description='The identifier of the DB cluster that this instance will belong to')
+    ] = None,
+    vpc_security_group_ids: Annotated[
+        Optional[List[str]], Field(description='A list of EC2 VPC security groups to associate with this DB instance')
+    ] = None,
+    availability_zone: Annotated[
+        Optional[str], Field(description='The Availability Zone where the DB instance will be created')
+    ] = None,
+    db_subnet_group_name: Annotated[
+        Optional[str], Field(description='A DB subnet group to associate with this DB instance')
+    ] = None,
+    multi_az: Annotated[
+        Optional[bool], Field(description='Specifies if the DB instance is a Multi-AZ deployment')
+    ] = False,
+    engine_version: Annotated[
+        Optional[str], Field(description='The version number of the database engine to use')
+    ] = None,
+    storage_type: Annotated[
+        Optional[str], Field(description='The storage type to be associated with the DB instance (standard, gp2, io1)')
+    ] = None,
+    storage_encrypted: Annotated[
+        Optional[bool], Field(description='Specifies whether the DB instance is encrypted')
+    ] = None,
+    port: Annotated[
+        Optional[int], Field(description='The port number on which the database accepts connections')
+    ] = None,
+    publicly_accessible: Annotated[
+        Optional[bool], Field(description='Specifies whether the DB instance is publicly accessible')
+    ] = None,
+    backup_retention_period: Annotated[
+        Optional[int], Field(description='The number of days for which automated backups are retained')
+    ] = None,
+    ctx: Context = None,
+) -> Dict[str, Any]:
+    """Create a new RDS database instance.
+
+    Args:
+        db_instance_identifier: The identifier for the DB instance
+        db_instance_class: The compute and memory capacity of the DB instance
+        engine: The name of the database engine to be used for this instance
+        allocated_storage: The amount of storage (in GiB) to be allocated
+        master_username: The name of the master user for the DB instance
+        master_user_password: The password for the master user
+        manage_master_user_password: Specifies whether to manage the master user password with AWS Secrets Manager
+        db_name: The name for your database
+        db_cluster_identifier: The identifier of the DB cluster that this instance will belong to
+        vpc_security_group_ids: A list of EC2 VPC security groups to associate with this DB instance
+        availability_zone: The Availability Zone where the DB instance will be created
+        db_subnet_group_name: A DB subnet group to associate with this DB instance
+        multi_az: Specifies if the DB instance is a Multi-AZ deployment
+        engine_version: The version number of the database engine to use
+        storage_type: The storage type to be associated with the DB instance
+        storage_encrypted: Specifies whether the DB instance is encrypted
+        port: The port number on which the database accepts connections
+        publicly_accessible: Specifies whether the DB instance is publicly accessible
+        backup_retention_period: The number of days for which automated backups are retained
+        ctx: MCP context for logging and state management
+
+    Returns:
+        Dict[str, Any]: The response from the AWS API
+    """
+    # Get RDS client
+    rds_client = RDSConnectionManager.get_connection()
+    
+    # Check if server is in readonly mode
+    if not check_readonly_mode('create', Context.readonly_mode(), ctx):
+        return {'error': ERROR_READONLY_MODE}
+
+    # validate identifier
+    if not validate_db_identifier(db_instance_identifier):
+        error_msg = ERROR_INVALID_PARAMS.format('db_instance_identifier must be 1-63 characters, begin with a letter, and contain only alphanumeric characters and hyphens')
+        if ctx:
+            await ctx.error(error_msg)
+        return {'error': error_msg}
+
+    try:
+        params = {
+            'DBInstanceIdentifier': db_instance_identifier,
+            'DBInstanceClass': db_instance_class,
+            'Engine': engine,
+        }
+
+        # Different parameter requirements based on whether this is a cluster instance
+        if db_cluster_identifier:
+            # Instance for existing cluster
+            params['DBClusterIdentifier'] = db_cluster_identifier
+        else:
+            # Standalone instance needs additional parameters
+            if allocated_storage is None:
+                error_msg = ERROR_INVALID_PARAMS.format('allocated_storage is required for standalone instances')
+                if ctx:
+                    await ctx.error(error_msg)
+                return {'error': error_msg}
+                
+            if master_username is None and not master_user_password and not manage_master_user_password:
+                error_msg = ERROR_INVALID_PARAMS.format('master_username and either master_user_password or manage_master_user_password are required for standalone instances')
+                if ctx:
+                    await ctx.error(error_msg)
+                return {'error': error_msg}
+                
+            params['AllocatedStorage'] = allocated_storage
+            
+            if master_username:
+                params['MasterUsername'] = master_username
+                
+            if master_user_password:
+                params['MasterUserPassword'] = master_user_password
+            elif manage_master_user_password:
+                params['ManageMasterUserPassword'] = manage_master_user_password
+
+        # Add optional parameters if provided
+        if db_name:
+            params['DBName'] = db_name
+        if vpc_security_group_ids:
+            params['VpcSecurityGroupIds'] = vpc_security_group_ids
+        if availability_zone:
+            params['AvailabilityZone'] = availability_zone
+        if db_subnet_group_name:
+            params['DBSubnetGroupName'] = db_subnet_group_name
+        if multi_az is not None:
+            params['MultiAZ'] = multi_az
+        if engine_version:
+            params['EngineVersion'] = engine_version
+        if storage_type:
+            params['StorageType'] = storage_type
+        if storage_encrypted is not None:
+            params['StorageEncrypted'] = storage_encrypted
+        if port is not None:
+            params['Port'] = port
+        elif not db_cluster_identifier:  # Don't set port for cluster instances
+            params['Port'] = get_engine_port(engine)
+        if publicly_accessible is not None:
+            params['PubliclyAccessible'] = publicly_accessible
+        if backup_retention_period is not None:
+            params['BackupRetentionPeriod'] = backup_retention_period
+
+        # MCP tags
+        params = add_mcp_tags(params)
+
+        logger.info(f"Creating DB instance {db_instance_identifier} with engine {engine}")
+        response = await asyncio.to_thread(rds_client.create_db_instance, **params)
+        logger.success(f"Successfully created DB instance {db_instance_identifier}")
+        
+        result = format_aws_response(response)
+        result['message'] = SUCCESS_CREATED.format(f'DB instance {db_instance_identifier}')
+        result['formatted_instance'] = format_instance_info(result.get('DBInstance', {}))
+        
+        return result
+    except Exception as e:
+        # The decorator will handle the exception
+        raise e
